+++ R/cv_document.R
@@ -3,11 +3,16 @@
 #' This output format provides support for including LaTeX dependencies and
 #' bibliography entries in extension of the bookdown::pdf_document2() format.
 #'
-#' @param ... Arguments passed to bookdown::pdf_document2()
+#' @inheritParams rmarkdown::pdf_document
+#' @param ... Arguments passed to bookdown::pdf_document2().
+#' @param pandoc_vars Pandoc variables to be passed to the template.
 #'
 #' @export
-cv_document <- function(...) {
-  config <- bookdown::pdf_document2(...)
+cv_document <- function(..., pandoc_args = NULL, pandoc_vars = NULL) {
+  for (i in seq_along(pandoc_vars)){
+    pandoc_args <- c(pandoc_args, rmarkdown::pandoc_variable_arg(names(pandoc_vars)[[i]], pandoc_vars[[i]]))
+  }
+  config <- bookdown::pdf_document2(..., pandoc_args = pandoc_args)
 
   pre <- config$pre_processor
   config$pre_processor <- function(metadata, input_file, runtime, knit_meta,