+++ R/formats.R
@@ -45,12 +45,33 @@ awesomecv <- function(...) {
 }
 
 #' @rdname cv_formats
+#'
+#' @param theme The theme used for the template.
+#'
 #' @export
-moderncv <- function(...) {
+moderncv <- function(..., theme = c("casual", "classic", "oldstyle", "banking", "fancy")) {
+  theme <- match.arg(theme)
   template <- system.file("rmarkdown", "templates", "moderncv",
     "resources", "moderncv.tex",
     package = "vitae"
   )
   copy_supporting_files("moderncv")
-  cv_document(..., template = template, citation_package = "biblatex", latex_engine = "xelatex")
+  cv_document(..., pandoc_vars = list(theme = theme),
+              template = template, citation_package = "biblatex", latex_engine = "pdflatex")
+}
+
+#' @rdname cv_formats
+#'
+#' @export
+latexcv <- function(..., theme = c("classic", "modern", "rows", "sidebar", "two_column")) {
+  theme <- match.arg(theme)
+  if(theme != "classic"){
+    stop("Only the classic theme is currently supported.")
+  }
+  template <- system.file("rmarkdown", "templates", "latexcv",
+                          "resources", theme, "main.tex",
+                          package = "vitae"
+  )
+  copy_supporting_files("latexcv")
+  cv_document(..., template = template, citation_package = "biblatex")
 }